fix: change dir structure
