+++ src/apis/tencent/error.ts
