+++ tsconfig.json
@@ -1,7 +1,6 @@
 {
   "compilerOptions": {
     "allowSyntheticDefaultImports": true,
-    "baseUrl": ".",
     "composite": true,
     "declaration": true,
     "declarationMap": true,
@@ -15,8 +14,9 @@
     "module": "commonjs",
     "moduleResolution": "node",
     "sourceMap": false,
+    "baseUrl": ".",
     "paths": {
-      "*": ["typings/*"]
+      "*": ["*", "src/*"]
     },
     "resolveJsonModule": true,
     "noUnusedLocals": true,
@@ -25,5 +25,5 @@
     "rootDir": "src",
     "outDir": "dist"
   },
-  "include": ["src", "src/typings/**/*.d.ts", "src/cli/migrate.ts"]
+  "include": ["src"]
 }
